refactor : 수동 로또 발매기 전략 패턴 추가
