+++ src/main/kotlin/domain/model/Lotto.kt
@@ -1,7 +1,7 @@
 package domain.model
 
 import ERROR
-import domain.strategy.KoreanLottoGenerator.Companion.LOTTO_SIZE
+import domain.strategy.AutoLottoGenerator.Companion.LOTTO_SIZE
 
 class Lotto(val numbers: List<LottoNumber>) {
     init {
