--- src/main/kotlin/domain/strategy/LottoCountry.kt
@@ -1,7 +1,0 @@
-package domain.strategy
-
-import domain.model.LottoNumber
-
-interface LottoCountry {
-    fun generateNumber(): List<LottoNumber>
-}
