+++ src/main/kotlin/domain/strategy/LottoGeneratorType.kt
@@ -1,0 +1,7 @@
+package domain.strategy
+
+import domain.model.Lotto
+
+interface LottoGeneratorType {
+    fun generateNumber(): List<Lotto>
+}
